Add a shared layout for wiki lists of links
Add GitHub and Visual Studio Code icons in the process and move
music.css to wiki/links.css not to be confused with links.css.
The links page will likely be integrated into sevaral wiki pages but is
confusing right now as two layouts are named links at the moment.